requirements.txt
flask>=2.2.0
pyrebase4==3.0.27
python-dotenv